requirements.txt
python-dotenv==1.0.0
pypdf2==3.0.1
openai==1.12.0
azure-search-documents==11.4.0
tiktoken==0.5.2
streamlit==1.31.0
pandas==2.2.0